semonitor: Do not log zero length messages in rec file
The code currently has code in place that only writes logging output when
a message is larger then 0. That's good, but this does not apply to the
rec file, which happily logs a zero length message. In turn, logging a recMsg,
even a zero length one, will always get the signature prepended.

This causes our rec files to be larger then they actually are. This
specifically happens when the message stream is initially incomplete.
Here validation will fail and thus produce an empty message.

This patch changes the logic in exiting the function when the message
is empty.

This can also be seen in the produced rec files in the testing directory.
Some start with a 'double header'. To ensure this is caught in the future
the test files have been updated by stripping the first 4 bytes where
applicable.

Finally, the test suite has been extended to also feed the rec files as
source to see if we can generate identical files.

+++ test/test.sh
@@ -26,5 +26,18 @@ do
     cat "test/json/${SAMPLE}.json" | conversion/se2csv.py -p "${TMPSE2CSV}/${SAMPLE}" -t
     diff "${TMP}/se2csv" "test/csv/${SAMPLE}/" -w
 
+    if ! ./semonitor.py -r "${TMP}/${SAMPLE}.re.rec" -o "${TMP}/${SAMPLE}.re.json" "test/rec/${SAMPLE}.rec" ${KEY_OPTION}; then
+        echo "Failed to run semonitor on rec file."
+        exit 1
+    fi
+    if ! cmp -l "${TMP}/${SAMPLE}.re.rec" "test/rec/${SAMPLE}.rec"; then
+        echo "Failed to compare re.rec file."
+        exit 1
+    fi
+    if ! diff "${TMP}/${SAMPLE}.re.json" "test/json/${SAMPLE}.json"; then
+        echo "Failed to compare re.JSON file."
+        exit 1
+    fi
+
     rm -rf "${TMP}"
 done
